Add authentication handlers and services
